(8.4) PS-9834: Fix calcuation of current audit logs file size
std::accumulate's return type is based on the type of the initial value
for the accumulator.  The integer literal - `0` - will be of type int,
typically a signed 32-bit value.  This results in an integer overflow bug
when the total size of audit logs exceeds 2GiB.

To fix this behavior, initialize explicitly with an unsigned long long.